+++ tests/IdeaStrike.Tests/AdminModuleTests/when_viewing_the_admin_page.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Nancy;
+﻿using Nancy;
 using Xunit;
