+++ tests/IdeaStrike.Tests/CommentModuleTests/when_adding_a_new_comment.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Ideastrike.Nancy.Models;
 using Moq;
 using Xunit;
